Conflicts:
	sql/auth/sql_user_table.cc
	sql/item_timefunc.cc
	sql/sp.cc
	sql/sql_cache.cc
	sql/sql_class.cc
	sql/sql_partition.cc
	sql/sql_table.cc
	sql/sql_update.cc
	sql/table.cc
	sql/trigger_loader.cc

+++ include/m_string.h
@@ -232,9 +232,9 @@ extern char *longlong10_to_str(longlong val,char *dst,int radix);
 #include <mysql/mysql_lex_string.h>
 typedef struct st_mysql_lex_string LEX_STRING;
 
-#define STRING_WITH_LEN(X) (X), ((size_t) (sizeof(X) - 1))
-#define USTRING_WITH_LEN(X) ((uchar*) X), ((size_t) (sizeof(X) - 1))
-#define C_STRING_WITH_LEN(X) ((char *) (X)), ((size_t) (sizeof(X) - 1))
+#define STRING_WITH_LEN(X) (X), ((sizeof(X) - 1))
+#define USTRING_WITH_LEN(X) ((uchar*) X), ((sizeof(X) - 1))
+#define C_STRING_WITH_LEN(X) ((char *) (X)), ((sizeof(X) - 1))
 
 struct st_mysql_const_lex_string
 {
